requirements.txt
flask==3.1.1
flask-cors==4.0.0
google-cloud-bigquery==3.17.2
python-dotenv==1.1.0
pandas==2.2.3
pyarrow==15.0.2
pandas-gbq==0.21.0